refactor(ensindexer): move ponder config creation into `src/ponder/config`
This will make it easier to read the ponder config object from other files.
